Update Readme.md to reflect less nonsense no one cares about.
ChatGPT vibe coding and lack of proofreading on my part can only go so far.  This typing project was a means to an end, made running on fumes, vaguely remembering the things I actually learned previously enough to know what I didn't want the robit to do.  The robit is so delighted to help though, and it only tries to please when it co-opts my fussy crap as being something I think is worthwhile to an audience of tens.